+++ src/templates/empty-points-list-template.js
@@ -1,0 +1,4 @@
+export const emptyPointsTemplate = (emptyPointTextValue) => `
+    <p class="trip-events__msg">
+      ${emptyPointTextValue}
+    </p>`;
