+++ src/templates/error-template.js
@@ -1,0 +1,4 @@
+export const errorTemplate = () => `
+      <p class="trip-events__msg">
+        Sorry, there was an error loading the data
+      </p>`;
