+++ src/templates/filter-template.js
@@ -1,0 +1,18 @@
+const createFilterItemTemplate = (filter, currentFilterType) => (
+  `<div class="trip-filters__filter">
+    <input id="filter-${filter.name}" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter"
+${filter.type === currentFilterType ? 'checked' : ''} value="${filter.type}" ${filter.count === 0 ? 'disabled' : ''}>
+    <label class="trip-filters__filter-label" for="filter-${filter.name}">${filter.name}</label>
+  </div>`
+);
+
+export const createFilterTemplate = (filterItems, currentFilterType) => {
+  const filterItemsTemplate = filterItems
+    .map((filter) => createFilterItemTemplate(filter, currentFilterType))
+    .join('');
+
+  return `<form class="trip-filters" action="#" method="get">
+    ${filterItemsTemplate}
+    <button class="visually-hidden" type="submit">Accept filter</button>
+    </form>`;
+};
