+++ src/templates/loading-template.js
@@ -1,0 +1,4 @@
+export const loadingTemplate = () => `
+      <p class="trip-events__msg">
+        Loading...
+      </p>`;
