+++ src/templates/menu-template.js
@@ -1,0 +1,4 @@
+export const menuTemplate = () => `
+      <nav class="trip-controls__trip-tabs  trip-tabs">
+        <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+      </nav>`;
