+++ src/templates/points-list-template.js
@@ -1,0 +1,3 @@
+export const pointsListTemplate = () => `
+      <ul class="trip-events__list">
+      </ul>`;
