+++ src/view/empty-points-list-view.js
@@ -1,11 +1,6 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {FILTER_TYPE} from '../const.js';
-
-const NoPointsTextType = {
-  [FILTER_TYPE.EVERYTHING]: 'Click New Event to create your first point',
-  [FILTER_TYPE.PAST]: 'There are no past events now',
-  [FILTER_TYPE.FUTURE]: 'There are no future events now',
-};
+import {EMPTY_POINTS_TEXT} from '../const.js';
+import {emptyPointsTemplate} from '../templates/empty-points-list-template.js';
 
 class EmptyPointsListView extends AbstractView {
   #filterType;
@@ -16,12 +11,9 @@ class EmptyPointsListView extends AbstractView {
   }
 
   get template() {
-    const emptyPointTextValue = NoPointsTextType[this.#filterType];
+    const emptyPointTextValue = EMPTY_POINTS_TEXT[this.#filterType];
 
-    return `
-      <p class="trip-events__msg">
-        ${emptyPointTextValue}
-      </p>`;
+    return emptyPointsTemplate(emptyPointTextValue);
   }
 }
