+++ src/view/error-view.js
@@ -1,11 +1,9 @@
 import AbstractView from '../framework/view/abstract-view.js';
+import {errorTemplate} from '../templates/error-template.js';
 
 class ErrorView extends AbstractView {
   get template() {
-    return `
-      <p class="trip-events__msg">
-        Sorry, there was an error loading the data
-      </p>`;
+    return errorTemplate;
   }
 }
