+++ src/view/menu-view.js
@@ -1,11 +1,9 @@
 import AbstractView from '../framework/view/abstract-view.js';
+import {menuTemplate} from '../templates/menu-template.js';
 
 class MenuView extends AbstractView {
   get template() {
-    return `
-      <nav class="trip-controls__trip-tabs  trip-tabs">
-        <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-      </nav>`;
+    return menuTemplate;
   }
 }
